Tweak landing page and add new page with all activities

+++ src/main/webapp/components/ActivityTable.js
@@ -9,8 +9,8 @@ import {formatKm} from "../utils/DistanceUtils";
 
 type ActivityTableProps = {
     activities: Array<Activity>,
-    editHandler: (Activity) => void,
-    deleteHandler: (Activity) => void
+    editHandler?: (Activity) => void,
+    deleteHandler?: (Activity) => void
 };
 
 export default class ActivityTable extends React.Component<ActivityTableProps> {
@@ -31,7 +31,7 @@ export default class ActivityTable extends React.Component<ActivityTableProps> {
                         <th>Duration</th>
                         <th>Distance</th>
                         <th>Split time (1 km)</th>
-                        <th>Actions</th>
+                        {this.props.editHandler && this.props.deleteHandler && <th>Actions</th>}
                     </tr>
                 </thead>
                 <tbody>
@@ -55,20 +55,26 @@ export default class ActivityTable extends React.Component<ActivityTableProps> {
             <td>{formatHourMinutes(a.duration)}</td>
             <td>{formatKm(a.distance)}</td>
             <td>{formatMinuteSeconds(1000 * a.duration / a.distance)}</td>
-            <td>
-                <DropdownButton bsSize="xsmall" title={<Glyphicon glyph="cog"/>} id="foo">
-                    <MenuItem eventKey="edit" onSelect={() => this.onEdit(a)}>Edit</MenuItem>
-                    <MenuItem eventKey="delete" onSelect={() => this.onDelete(a)}>Delete</MenuItem>
-                </DropdownButton>
-            </td>
+            {this.props.editHandler && this.props.deleteHandler && this.renderActions(a)}
         </tr>;
     }
 
+    renderActions(activity: Activity) {
+        return (
+            <td>
+                <DropdownButton bsSize="xsmall" title={<Glyphicon glyph="cog"/>} id="foo">
+                    <MenuItem eventKey="edit" onSelect={() => this.onEdit(activity)}>Edit</MenuItem>
+                    <MenuItem eventKey="delete" onSelect={() => this.onDelete(activity)}>Delete</MenuItem>
+                </DropdownButton>
+            </td>
+        );
+    }
+
     onEdit(a: Activity) {
-        this.props.editHandler(a);
+        this.props.editHandler && this.props.editHandler(a);
     }
 
     onDelete(a: Activity) {
-        this.props.deleteHandler(a);
+        this.props.deleteHandler && this.props.deleteHandler(a);
     }
 }
